optimized before_script.sh for multiple executions

+++ before_script.sh
@@ -1,16 +1,35 @@
-echo "Getting composer"
-curl -O http://getcomposer.org/composer.phar
+serverUrl='http://127.0.0.1:4444'
+serverVersion='2.32.0'
+serverFile=selenium-server-standalone-$serverVersion.jar
+
+
+if [ ! -f composer.phar ]; then
+    echo "Getting composer"
+    curl -O http://getcomposer.org/composer.phar
+else
+    php composer.phar self-update
+fi
+
 echo "Dependencies"
-php composer.phar install --dev
+if [ ! -d vendor ] || [ ! -f vendor/autoload.php ]; then 
+    php composer.phar install --dev
+else
+    php composer.phar update --dev
+fi
 echo "Starting Python HTTP server"
 cd selenium-1-tests
 python -m SimpleHTTPServer 8080 > /dev/null 2>&1 &
 cd ..
 echo "Starting xvfb"
-sh -e /etc/init.d/xvfb start
-export DISPLAY=:99.0
 echo "Starting Selenium"
-wget http://selenium.googlecode.com/files/selenium-server-standalone-2.32.0.jar
-java -jar selenium-server-standalone-2.32.0.jar > /dev/null 2>&1 &
-sleep 30
-echo "Finished setup"
+if [ ! -f $serverFile ]; then
+    wget http://selenium.googlecode.com/files/selenium-server-standalone-$serverVersion.jar -O $serverFile
+fi
+xvfb-run java -jar $serverFile > /dev/null 2>&1 &
+
+wget --retry-connrefused --tries=60 --waitretry=1 --output-file=/dev/null $serverUrl/wd/hub/status -O /dev/null
+if [ ! $? -eq 0 ]; then
+    echo "Selenium Server not started"
+else
+    echo "Finished setup"
+fi
